dataset/code-template/requirements.txt
scipy
pillow
torch
torchvision
